INTYGFV-13859: Refactored CertificateRelation(s) to use Lombok and moved the classes to subpackage metadata.
